+++ source/darefl/CMakeLists.txt
@@ -7,7 +7,7 @@ add_subdirectory(model)
 add_subdirectory(resources)
 add_subdirectory(famousloader)
 add_subdirectory(grapheditor)
-add_subdirectory(importdatawidgets)
+add_subdirectory(importdataview)
 add_subdirectory(layereditor)
 add_subdirectory(materialeditor)
 add_subdirectory(quicksimeditor)
