+++ source/darefl/importdataview/datarowstrategy.h
